refactor(components): fix some routing in App
