+++ src/components/App/App.jsx
@@ -1,29 +1,17 @@
 // import { Routes, Route } from 'react-router-dom';
-import { useDispatch } from 'react-redux';
-import { useEffect } from 'react';
-
-// import Feedback from '../Feedback/Feedback';
-// import HowServiceWorks from '../HowServiceWorks/HowServiceWorks';
-// import AboutService from '../AboutService/AboutService';
-// import BestService from '../BestService/BestService';
-// import Modal from '../Modal/Modal';
-import Footer from '../Footer/Footer';
-// import Login from '../Login/Login';
-// import Registration from '../Registration/Registration';
-// import Profile from '../Profile/Profile';
-// import ProtectedRoute from '../ProtectedRoute/ProtectedRoute';
 import '../../index.css';
 
-import AboutUs from '../AboutUs/AboutUs';
-import BestService from '../BestService/BestService';
-import Feedback from '../Feedback/Feedback';
-import Header from '../Header/Header';
-import HowServiceWorks from '../HowServiceWorks/HowServiceWorks';
-import MainScreen from '../MainScreen/MainScreen';
-import { fetchAutoServices } from '../../store/autoServicesSlice';
-import Search from '../Search/Search';
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { Route, Routes } from 'react-router';
+
+import MainPage from '../../Pages/MainPage/MainPage';
 // import Ymap from '../Ymap/Ymap';
+import { fetchAutoServices } from '../../store/autoServicesSlice';
+import Footer from '../Footer/Footer';
+import Header from '../Header/Header';
 import style from './styles/App.module.css';
+import ServicePage from '../../Pages/ServicePage/ServicePages';
 
 function App() {
   // const accessToken = getCookie('accessToken').replace('Bearer', '');
@@ -42,12 +30,10 @@ function App() {
   return (
     <div className={style.app}>
       <Header />
-      <MainScreen />
-      <Search />
-      <BestService />
-      <AboutUs />
-      <HowServiceWorks />
-      <Feedback />
+      <Routes>
+        <Route path="/" element={<MainPage />} />
+        <Route path="/service/:id" element={<ServicePage />} />
+      </Routes>
       <Footer />
       {/* <Routes>
         <Route path="/" element={<MainPage />} />
